Add Testing Files for add and remove modules
