+++ src/__mocks__/storage.js
@@ -1,0 +1,8 @@
+let holder = [];
+function saveStorage(items) {
+  holder = items;
+}
+function getStorage() {
+  return holder;
+}
+export { saveStorage, getStorage };
