+++ src/addlist.js
@@ -1,5 +1,7 @@
+/* eslint-disable import/no-cycle */
 import { saveStorage, getStorage } from './storage.js';
-import populateList from './populateList.js';
+// const { getStorage, saveStorage } = storage;
+import populateList from './index.js';
 
 const addNewTask = (input) => {
   const taskList = getStorage();
@@ -12,9 +14,12 @@ const addNewTask = (input) => {
   taskList.push(task);
 
   saveStorage(taskList);
-  input.value = '';
-  populateList();
+  // input.value = '';
+  populateList(taskList);
   return task;
+  // if (input.value === '') {
+  //   return;
+  // }
 };
 
 export default addNewTask;
