+++ src/completed.js
@@ -1,3 +1,4 @@
+/* eslint-disable import/no-cycle */
 import { saveStorage, getStorage } from './storage.js';
 
 const trashCompleted = () => {
