+++ src/trashTask.js
